backend/pyproject.toml
[tool.poetry]
name = "my-fastapi-project"
version = "0.1.0"
description = ""
authors = ["shotarai <second.arasho@outlook.jp>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.112.1"
uvicorn = "^0.30.6"
ruamel-yaml = "^0.18.6"
google-generativeai = "^0.7.2"
python-dotenv = "^1.0.1"


[tool.poetry.group.dev.dependencies]
isort = "^5.13.2"
black = "^24.8.0"
mypy = "^1.11.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
